fix :gradle 수정

+++ stock-api/src/main/java/com/hyun/stockapi/stock_api/dto/StockResponseDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class StockResponseDto {
     private String companyName;
     private String tradeDate;
-    private long closingPrice;
+    private float closingPrice;
 
 
 }
